Fix: Handle calendar misses
It has been noticed that Airbnb will sometimes return an empty calendar
which causes problems if the existing calendar only has a single event
either from a long stay or just no upcoming reservations after the
current active one. This change adds causes a configuration to keep the
current calendar until a set number of misses (currently 2) have
occurred. This should help smooth over these temporary issues with
Airbnb's calendar API. There is currently no user configuration for this
value, but that could be added in the future if needed.

Signed-off-by: Andrew Grimberg <tykeal@bardicgrove.org>